Modified splitmerge to iterate only over pairs of neighbors which are correlated. It also uses a set() object to contain these neighbor pairs, in order to reduce duplicates. For the test network of 52 stations, this reduces the number of pairs from 1378 to 673.
Experimented with different methods for finding the median in the slopes generated by the Kendall-Theil method in minbic(). For large segments with 900 data values, we compute ~400,000 slopes. The naive way of computing the median - where you sort them and then find the middle indexed value - is slow due to the sorting step. I found a discussion thread on the NumPy implementation of the median-finding method, which is based on a version Hoare's algorthim modified by Niklaus Wirth. I converted this method to pure Python and added it to code.util(). This dramatically speeds up the median-computation process for the KTHSLR1 and KTHTPR1 methods.