Minor tweaks to chgptmodels. I haven't gone through the log of code review changes; that will happen this weekend.
Experimented with parallelizing splitmerge(). Essentially, I'm calling minbic() many times. So, I implemented a way to handle this with the Python libraries threading and multiprocessing. Threading doesn't work well (GIL issues, so we're still running in serial). Multiprocessing on the other hand works fantastic! By pooling all the minbic calculations for each pair of stations and letting them run asynchronously, I speed up the entire splitmerge process by 4x (not surprising because I'm running a Pentium i5). It's hackish code for now, but the only issue is that since the processes run asynchronously, all the terminal stuff gets FUBAR. This isn't necessarily a problem; at some future time logging will be implemented, and I can just have minbic record its output as a string whcih is appended to logging at the end. The minbic debug statements will be out of order, but that doesn't matter - the summary at the end (MD TESTSEG ADJ) is the important sanity check.

Experimented with unconfounding the changepoints in test_drive. I've saved one of the sped-up splitmerge results to disk (it's a pickled dictionary) so that I can analyze the detected changepoints. With numpy, I think I can do the entire unconfounding process - that is, figuring from all the breakpoints and the stations which contribute them, what the *actual* inhomogeneities are in each station's data - realtively easily. If it works with numpy, I'll leave it in there and finish up the algorithm, and at some future version replace the numpy code with pure python (I'll just replace np.arrays with matrices - lists-of-lists).