Modified splitmerge to accept a list of station pairs; this delegates the work back to the top-level script which processes all of the input stations.
Added 2nd filter in unconfounding process. This is really just a stub; it tries to "absorb" undocumented changepoints to nearby known changepoints from a history or metadatafile. That's a version +1 function, so the basic functionality of finding windows around changepoints is implemented and hte rest can be fleshed out laters.

Added first attempt at 3rd filter. This compares nearby undocumented changepoints to condense the set a bit. Works pretty well actually; biggest problme I have is that I'm still running on a slightly different set of changepoints. For instance, on the test station network I've been using, the Fortran code yields 618 undocumented changepoints at this point in the code. My code yields 718. I've documented a few issues that contribute to this, but I will visit them in the future.